+++ crates/maple_3d/src/lib.rs
@@ -3,3 +3,16 @@ pub mod gltf;
 pub mod nodes;
 pub mod plugin;
 pub mod render_passes;
+
+pub mod prelude {
+    pub use crate::nodes::{
+        camera::{Camera3D, Camera3DBuilder},
+        directional_light::{DirectionalLight, DirectionalLightBuilder},
+        mesh::{Mesh3D, Mesh3DBuilder},
+        point_light::{PointLight, PointLightBuilder},
+    };
+
+    pub use crate::components::material::*;
+    pub use crate::gltf::GLTFLoader;
+    pub use crate::plugin::Core3D;
+}
