+++ crates/maple_engine/src/components/mod.rs
@@ -1,6 +1,6 @@
 //! Contains components that nodes use such as their transform or Mesh.
 
-pub mod event_reciever;
+mod event_reciever;
 // pub mod mesh;
 pub mod node_transform;
 
@@ -8,3 +8,5 @@ pub mod node_transform;
 pub use event_reciever::{EventReceiver, FixedUpdate, Ready, Update};
 // pub use mesh::Mesh;
 pub use node_transform::NodeTransform;
+
+pub use event_reciever::*;
