+++ crates/maple_engine/src/lib.rs
@@ -11,7 +11,9 @@ pub use scene::{Scene, SceneBuilder};
 pub use nodes::{Buildable, Builder, Node};
 
 pub mod prelude {
-    pub use crate::components::{NodeTransform, event_reciever::EventReceiver};
+    pub use crate::components::{
+        EventLabel, EventReceiver, FixedUpdate, NodeTransform, Ready, Update,
+    };
 
     pub use crate::input::*;
