+++ crates/maple_physics/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2024"
 [dependencies]
 nalgebra = {version = "0.34.1", features = ["convert-glam030"]}
 rapier3d = "0.30.1"
-maple = {version = "0.3.0-dev", path = "../../"}
-
+maple_engine = {path = "../maple_engine", version = "0.3.0-dev"}
+maple_app = {path = "../maple_app", version = "0.3.0-dev"}
+glam = "0.30.9"
