+++ crates/maple_physics/src/lib.rs
@@ -1,3 +1,9 @@
 pub mod nodes;
 pub mod plugin;
 pub mod resource;
+
+pub mod prelude {
+    pub use crate::nodes::*;
+    pub use crate::plugin::Physics3D;
+    pub use crate::resource::Physics;
+}
