+++ crates/maple_physics/src/nodes/collider.rs
@@ -1,10 +1,7 @@
-use maple::{
-    derive::Node,
-    engine::{
-        Scene,
-        nodes::node_builder::{Buildable, Builder, NodePrototype},
-    },
-    math::Vec3,
+use glam::Vec3;
+use maple_engine::{
+    Buildable, Builder, Node, Scene,
+    nodes::node_builder::NodePrototype,
     prelude::{EventReceiver, NodeTransform},
 };
 use rapier3d::prelude::{ActiveEvents, ColliderBuilder, ColliderHandle, Group, InteractionGroups};
@@ -45,13 +42,9 @@ pub enum CapsuleAxis {
     Z,
 }
 
-#[derive(Node)]
 pub struct Collider3D {
-    #[events]
     events: EventReceiver,
-    #[children]
     children: Scene,
-    #[transform]
     transform: NodeTransform,
 
     pub(crate) handle: Option<ColliderHandle>,
@@ -68,6 +61,24 @@ pub struct Collider3D {
     contact_skin: f32,
     enabled: bool,
     active_events: ActiveEvents,
+}
+
+impl Node for Collider3D {
+    fn get_transform(&mut self) -> &mut NodeTransform {
+        &mut self.transform
+    }
+
+    fn get_children_mut(&mut self) -> &mut Scene {
+        &mut self.children
+    }
+
+    fn get_children(&self) -> &Scene {
+        &self.children
+    }
+
+    fn get_events(&mut self) -> &mut EventReceiver {
+        &mut self.events
+    }
 }
 
 impl Collider3D {
@@ -288,8 +299,12 @@ impl Collider3DBuilder {
     }
 
     /// Create a triangle collider
-    pub fn triangle(a: Vec3, b: Vec3, c: Vec3) -> Self {
-        Collider3D::builder().shape(ColliderShape::Triangle { a, b, c })
+    pub fn triangle(a: impl Into<Vec3>, b: impl Into<Vec3>, c: impl Into<Vec3>) -> Self {
+        Collider3D::builder().shape(ColliderShape::Triangle {
+            a: a.into(),
+            b: b.into(),
+            c: c.into(),
+        })
     }
 
     /// Set the collider shape
