+++ examples/mandelbrot/main.rs
@@ -150,7 +150,7 @@ impl RenderNode for ShowPass {
         render_ctx.render(node_ctx, |mut fb| {
             fb.bind_vertex_buffer(self.vertex_buffer.as_ref().unwrap())
                 .bind_descriptor_set(0, set)
-                .draw();
+                .draw_vertices();
         });
     }
 }
